Refactored HttpRequestPredicates  (#2412)
