+++ tests/ga/test_send_telemetry_events.py
@@ -42,7 +42,8 @@ from azurelinuxagent.common.version import CURRENT_VERSION, DISTRO_NAME, DISTRO_
 from azurelinuxagent.ga.collect_telemetry_events import _CollectAndEnqueueEvents
 from azurelinuxagent.ga.send_telemetry_events import get_send_telemetry_events_handler
 from tests.ga.test_monitor import random_generator
-from tests.protocol.mocks import MockHttpResponse, mock_wire_protocol, HttpRequestPredicates
+from tests.protocol.mocks import MockHttpResponse, mock_wire_protocol
+from tests.protocol.HttpRequestPredicates import HttpRequestPredicates
 from tests.protocol.mockwiredata import DATA_FILE
 from tests.tools import AgentTestCase, clear_singleton_instances, mock_sleep
 from tests.utils.event_logger_tools import EventLoggerTools
