+++ tests/protocol/mockwiredata.py
@@ -18,6 +18,7 @@
 import re
 
 from azurelinuxagent.common.utils.textutil import parse_doc, find, findall
+from tests.protocol.HttpRequestPredicates import HttpRequestPredicates
 from tests.tools import load_bin_data, load_data, MagicMock, Mock
 from azurelinuxagent.common.exception import HttpError, ResourceGoneError
 from azurelinuxagent.common.future import httpclient
@@ -118,10 +119,6 @@ DATA_FILE_VM_SETTINGS["in_vm_artifacts_profile"] = "hostgaplugin/in_vm_artifacts
 
 DATA_FILE_STATUS_BLOB = DATA_FILE.copy()
 DATA_FILE_STATUS_BLOB["ext_conf"] = "wire/ext_conf_mock_status_blob.xml"
-
-
-def is_ga_manifest_request(url):
-    return "manifest_of_ga.xml" in url
 
 
 class WireProtocolData(object):
@@ -260,7 +257,7 @@ class WireProtocolData(object):
             if "manifest.xml" in url:
                 content = self.manifest
                 self.call_counts["manifest.xml"] += 1
-            elif is_ga_manifest_request(url):
+            elif HttpRequestPredicates.is_ga_manifest_request(url):
                 content = self.ga_manifest
                 self.call_counts["manifest_of_ga.xml"] += 1
             elif "ExampleHandlerLinux" in url:
